Consolidate imports with index.ts
