+++ src/managers/index.ts
@@ -1,0 +1,3 @@
+export * from './configuration';
+export * from './symbol';
+export * from './workspace';
