+++ src/providers/index.ts
@@ -1,0 +1,15 @@
+export * from './codelens';
+export * from './completion';
+export * from './debug-configuration';
+export * from './definition';
+export * from './document-highlight';
+export * from './document-link';
+export * from './document-semantic-tokens';
+export * from './document-symbol';
+export * from './folding-range';
+export * from './hover';
+export * from './implementation';
+export * from './reference';
+export * from './rename';
+export * from './task';
+export * from './workspace-symbol';
